Create Ardpi_Serial_test
Program that tests the serial connection between the Arduino UNO board and the Raspberry PI B+. When configured correctly, Rasperry PI will send a byte. When it sends a 1 it turns on an LED on the Arduino. When a 0 is sent the LED is turned off. Anything else does not effect the LED, but in all cases the byte is sent back to the Pi to confirm data received.